+++ src/main/java/kr/hhplus/be/server/queue/domain/QueueToken.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.domain.queue;
+package kr.hhplus.be.server.queue.domain;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
